requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
streamlit==1.54.0
sentence-transformers==2.2.2
transformers==4.33.2
torch==2.10.0
faiss-cpu==1.13.2
langdetect==1.0.9
googletrans==4.0.0-rc1
numpy==1.25.1
pandas==2.3.3